+++ pm/drivers/psc/psc.c
@@ -229,7 +229,7 @@ void psc_pd_wait(struct device *dev, struct psc_pd *pd)
 			/* Directly convert to psc to get psc_idx */
 			pm_trace(TRACE_PM_ACTION_PSC_TRANSITION_TIMEOUT | TRACE_PM_ACTION_FAIL,
 				 (u32) (((u32) ((to_psc_drv_data(get_drv_data(dev)))->psc_idx) << TRACE_PM_VAL_PSC_SHIFT) |
-				 ((u16) psc_pd_idx(dev, pd) << TRACE_PM_VAL_PD_SHIFT) | TRACE_PM_VAL_PD_POS1));
+				 ((u32) psc_pd_idx(dev, pd) << TRACE_PM_VAL_PD_SHIFT) | TRACE_PM_VAL_PD_POS1));
 		}
 	}
 }
@@ -280,7 +280,7 @@ void psc_pd_get(struct device *dev, struct psc_pd *pd)
 	pm_trace(TRACE_PM_ACTION_PD_GET,
 		 ((u32) psc->psc_idx << TRACE_PM_VAL_PSC_SHIFT) |
 		 (idx << TRACE_PM_VAL_PD_SHIFT) |
-		 (pd->use_count & TRACE_PM_VAL_MAX_PSC_DATA));
+		 ((u32)pd->use_count & TRACE_PM_VAL_MAX_PSC_DATA));
 
 	if ((pd->use_count++) != 0U) {
 		/* Nothing to do */
@@ -388,7 +388,7 @@ void psc_pd_put(struct device *dev, struct psc_pd *pd)
 	pm_trace(TRACE_PM_ACTION_PD_PUT,
 		 ((u32) psc->psc_idx << TRACE_PM_VAL_PSC_SHIFT) |
 		 (idx << TRACE_PM_VAL_PD_SHIFT) |
-		 (pd->use_count & TRACE_PM_VAL_MAX_PSC_DATA));
+		 ((u32)pd->use_count & TRACE_PM_VAL_MAX_PSC_DATA));
 
 	if ((--pd->use_count) != 0U) {
 		/* Nothing to do */
@@ -642,7 +642,7 @@ static void lpsc_module_sync_state(struct device	*dev,
 		if (!depends_dev) {
 			pm_trace(TRACE_PM_ACTION_PSC_INVALID_DEP_DATA | TRACE_PM_ACTION_FAIL,
 				 (u32) (((u32) psc->psc_idx << TRACE_PM_VAL_PSC_SHIFT) |
-				 ((u16) data->depends_psc_idx << TRACE_PM_VAL_PD_SHIFT) | TRACE_PM_VAL_PD_POS1));
+				 ((u32) data->depends_psc_idx << TRACE_PM_VAL_PD_SHIFT) | TRACE_PM_VAL_PD_POS1));
 		} else {
 			/*
 			 * Moving from a clock stop state to a clock enabled
@@ -717,7 +717,7 @@ static void lpsc_module_sync_state(struct device	*dev,
 		if (!depends_dev) {
 			pm_trace(TRACE_PM_ACTION_PSC_INVALID_DEP_DATA | TRACE_PM_ACTION_FAIL,
 				 (u32) (((u32) psc->psc_idx << TRACE_PM_VAL_PSC_SHIFT) |
-				 ((u16) data->depends_psc_idx << TRACE_PM_VAL_PD_SHIFT) | TRACE_PM_VAL_PD_POS2));
+				 ((u32) data->depends_psc_idx << TRACE_PM_VAL_PD_SHIFT) | TRACE_PM_VAL_PD_POS2));
 		} else if (domain_reset && (depends_dev == dev)) {
 			/* Ignore self dependencies during domain reset */
 		} else {
@@ -990,8 +990,8 @@ static void lpsc_module_get_internal(struct device *dev,
 	if (use) {
 		pm_trace(TRACE_PM_ACTION_MODULE_GET,
 			 ((u32) psc->psc_idx << TRACE_PM_VAL_PSC_SHIFT) |
-			 ((u16) lpsc_module_idx(dev, module) << TRACE_PM_VAL_LPSC_SHIFT) |
-			 (module->use_count & TRACE_PM_VAL_MAX_PSC_DATA));
+			 ((u32) lpsc_module_idx(dev, module) << TRACE_PM_VAL_LPSC_SHIFT) |
+			 ((u32)module->use_count & TRACE_PM_VAL_MAX_PSC_DATA));
 		module->use_count++;
 		if (module->use_count == 1U) {
 			lpsc_module_clk_get(dev, module);
@@ -1002,8 +1002,8 @@ static void lpsc_module_get_internal(struct device *dev,
 	if (ret) {
 		pm_trace(TRACE_PM_ACTION_RETENTION_GET,
 			 (u32) (((u32) psc->psc_idx << TRACE_PM_VAL_PSC_SHIFT) |
-			 ((u16) lpsc_module_idx(dev, module) << TRACE_PM_VAL_LPSC_SHIFT) |
-			 (module->ret_count & TRACE_PM_VAL_MAX_PSC_DATA)));
+			 ((u32) lpsc_module_idx(dev, module) << TRACE_PM_VAL_LPSC_SHIFT) |
+			 ((u32)module->ret_count & TRACE_PM_VAL_MAX_PSC_DATA)));
 		module->ret_count++;
 		if (module->ret_count == 1U) {
 			lpsc_module_clk_get(dev, module);
@@ -1026,8 +1026,8 @@ static void lpsc_module_put_internal(struct device *dev,
 	if (use) {
 		pm_trace(TRACE_PM_ACTION_MODULE_PUT,
 			 (u32) (((u32) psc->psc_idx << TRACE_PM_VAL_PSC_SHIFT) |
-			 ((u16) lpsc_module_idx(dev, module) << TRACE_PM_VAL_LPSC_SHIFT) |
-			 (module->use_count & TRACE_PM_VAL_MAX_PSC_DATA)));
+			 ((u32) lpsc_module_idx(dev, module) << TRACE_PM_VAL_LPSC_SHIFT) |
+			 ((u32)module->use_count & TRACE_PM_VAL_MAX_PSC_DATA)));
 		module->use_count--;
 		if (module->use_count == 0U) {
 			modify = STRUE;
@@ -1037,8 +1037,8 @@ static void lpsc_module_put_internal(struct device *dev,
 	if (ret) {
 		pm_trace(TRACE_PM_ACTION_RETENTION_PUT,
 			 (u32) (((u32) psc->psc_idx << TRACE_PM_VAL_PSC_SHIFT) |
-			 ((u16) lpsc_module_idx(dev, module) << TRACE_PM_VAL_LPSC_SHIFT) |
-			 (module->ret_count & TRACE_PM_VAL_MAX_PSC_DATA)));
+			 ((u32) lpsc_module_idx(dev, module) << TRACE_PM_VAL_LPSC_SHIFT) |
+			 ((u32)module->ret_count & TRACE_PM_VAL_MAX_PSC_DATA)));
 		module->ret_count--;
 		if (module->ret_count == 0U) {
 			modify = STRUE;
@@ -1160,7 +1160,7 @@ static s32 psc_initialize_pds(struct device *dev)
 
 		pm_trace(TRACE_PM_ACTION_PD_INIT,
 			 (u32) (((u32) psc->psc_idx << TRACE_PM_VAL_PSC_SHIFT) |
-			 ((u16) idx << TRACE_PM_VAL_PD_SHIFT)));
+			 ((u32) idx << TRACE_PM_VAL_PD_SHIFT)));
 		psc_pd_wait(dev, pd);
 		state = (u8) (psc_read(dev, PSC_PDSTAT((u32) idx)) & PDSTAT_STATE_MASK);
 
@@ -1275,7 +1275,7 @@ static s32 psc_initialize_modules(struct device *dev)
 		if (i == 0) {
 			pm_trace(TRACE_PM_ACTION_PSC_TRANSITION_TIMEOUT | TRACE_PM_ACTION_FAIL,
 				 (u32) ((((u32) psc->psc_idx) << TRACE_PM_VAL_PSC_SHIFT) |
-				 ((u16) idx << TRACE_PM_VAL_PD_SHIFT) | TRACE_PM_VAL_PD_POS3));
+				 ((u32) idx << TRACE_PM_VAL_PD_SHIFT) | TRACE_PM_VAL_PD_POS3));
 		}
 
 		v = psc_read(dev, PSC_MDSTAT((u32) idx));
