+++ rm/core/rm_core.c
@@ -342,7 +342,7 @@ s32 rm_core_get_resource_range(u32	*msg_recv,
 			TISCI_MSG_VALUE_RM_UNUSED_SECONDARY_HOST) ?
 		       msg->hdr.host :
 		       msg->secondary_host;
-		utype = RESASG_UTYPE(msg->type, msg->subtype);
+		utype = RESASG_UTYPE(msg->type, (u16)msg->subtype);
 		type_found = SFALSE;
 
 		rm_trace_sub(trace_action,
